refactor(common-util,ecommerce,payment,admin,batch): 금액 double→BigDecimal 전환(ADR-0006) + 리뷰 후속 5건
금액(상품가·소계·할인·세금·합계·결제·환불·원장·장바구니 합계)을 6개 모듈에서 double→BigDecimal
(scale 2/HALF_UP)로 전사 전환해 부동소수 부채를 근본 해결(B3). 비율(Coupon.discountRate)은 무차원
계수라 double 유지(곱셈 지점에서 BigDecimal.valueOf 무손실 변환). 반올림 경계 불변(값 보존, 표현만 정밀).
응답 JSON 표현 변화(100.0→100.00)는 ADR-0006에 기록된 의도된 예외. ADR-0003을 대체.

리뷰 후속 5건:
- 🔴 DB 마이그레이션 산출물: 멱등 ALTER TABLE 스크립트 db/migration/ 커밋(ecommerce·payment) + README.
  ddl-auto:update가 DOUBLE→DECIMAL을 미전환하므로 수동 1회 실행(Flyway 없음). H2 2.3.232로 실행 검증.
- 🟡 테스트 컨벤션: AdminRefundControllerTest의 scale 민감 eq(BigDecimal) → argThat(compareTo==0).
- 🟡 scale=2 drift 제거: MoneyUtils.MONEY_SCALE/MONEY_PRECISION public 상수화, 엔티티 9개 @Column이 참조.
- 🟢 nit: PricingService 할인 이중 반올림 제거(tax와 대칭) / MoneyUtils.format() locale 비의존
  (toPlainString) / ProductService.create price 입력단 scale 정규화.

테스트 79개 green.

+++ legacy-ecommerce/admin/src/main/java/com/legacy/shop/admin/client/ShopGateway.java
@@ -5,6 +5,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 /**
@@ -43,7 +44,7 @@ public class ShopGateway {
         return restTemplate.getForObject(ecommerceUrl + "/api/orders/" + id, Map.class);
     }
 
-    public Object refund(Long paymentId, double amount, String reason) {
+    public Object refund(Long paymentId, BigDecimal amount, String reason) {
         PaymentRefundRequest req = new PaymentRefundRequest(paymentId, amount, reason);
         return restTemplate.postForObject(paymentUrl + "/api/payments/refund", req, Map.class);
     }
