+++ legacy-ecommerce/ecommerce-service/src/main/java/com/legacy/shop/ecommerce/dto/CartItemResponse.java
@@ -1,9 +1,11 @@
 package com.legacy.shop.ecommerce.dto;
 
+import java.math.BigDecimal;
+
 public record CartItemResponse(
         Long productId,
         int quantity,
-        double unitPrice,
-        double lineTotal
+        BigDecimal unitPrice,
+        BigDecimal lineTotal
 ) {
 }
