+++ legacy-ecommerce/ecommerce-service/src/main/java/com/legacy/shop/ecommerce/dto/OrderResponse.java
@@ -1,5 +1,6 @@
 package com.legacy.shop.ecommerce.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -7,10 +8,10 @@ public record OrderResponse(
         Long id,
         Long customerId,
         List<OrderItemResponse> items,
-        double subtotal,
-        double discountAmount,
-        double tax,
-        double totalAmount,
+        BigDecimal subtotal,
+        BigDecimal discountAmount,
+        BigDecimal tax,
+        BigDecimal totalAmount,
         String status,
         LocalDateTime orderedAt,
         Long paymentId
